Fix check for g:rust_conceal_mod_path in after/syntax/rust.vim
The previous check was quite broken. If `g:rust_conceal_mod_path` wasn't
set, it would throw an error. If it was set to `0`, it would incorrectly
skip linking the highlight groups. It only worked properly for the case
where `g:rust_conceal_mod_path` was set to a non-zero value.