Support rustfmt 0.5.0 exit codes.
This change updates the acceptable exit codes from `rustfmt` when
deciding whether or not to write back the file or display the errors
location list.

As of rust-lang-nursery/rustfmt#923, `rustfmt` exits with either a 0, 1,
2, or 3 depending on what happened. Prior to this upstream change,
`rustfmt` would exit 0 when issues that cannot be automatically resolved
would occur. Now, the same errors (such as "line exceeded maximum
length") result in an exit of 3 (see the [README](line exceeded maximum
length) for more details).

This change will treat an exit of 3 as "okay" for writing back as
`rustfmt` didn't fail (it just couldn't auto-format a nicer solution).

Closes #68
Closes #75
References rust-lang-nursery/rustfmt#923