Don't fix paths after running cargo.
In Rust 1.12, Cargo now emits paths that are relative to the current
directory, not Cargo.toml, so the path rewriting is no longer necessary.

In some repository layouts with multiple crates in a workspace, relative
paths would be rewritten relative to the wrong crate's Cargo.toml.

Also remove the documentation for the g:cargo_manifest_name option which
isn't used any more.