Change `unsafe` highlighting to `Exception`
It has a new group now, `rustUnsafeKeyword`, which allows the user to
easily customise the highlighting if they desire. (Example: if you
wanted to visually forbid `unsafe`, you could highlight it as `Error` by
putting `hi link rustUnsafeKeyword Error` in your `.vimrc` or
`.vim/after/syntax/rust.vim`.)

I also made this group link to `Exception` rather than `Keyword`.
This is potentially mildly controversial. It’s common for `Exception`
and `Keyword` to be highlighted the same, so many users won’t see this
change. Still, it *is* a change, and one people may not like. We could
back that part of it out easily enough.

Closes #144.