Reinstate matchit support, but without loading it
We used to load matchit, but that got removed when we landed rust.vim in
the upstream Vim runtime. I am reinstating the matchpairs/b:match_skip
parts of it, but not the loading of matchit.vim. This seems to be the
convention among syntax files: support it, but don’t load it.

I recommend that users add this to their vimrc file (assuming Vim 8):

    packadd! matchit

Fixes #160.