Add a Travis CI testing suite infrastructure and two tests
The added scripts utilize a Docker container that has multiple versions
of Vim and Neovim, plus a latest install of Rust, plus an install
of vader.vim.

For reference, these are the repos I used to create the Docker image
currently hosted in hub.docker.com:

    A modified vim-testbed (Based on Fedora instead of Alpine Linux)

	https://github.com/da-x/vim-testbed/tree/rust.vim

    Vader.vim snapshot:

	https://github.com/da-x/vader.vim/tree/v2017-12-26

The execution of the test roots around test/run-tests, which invokes
Docker on itself, then executes Vim to drive the Vader-based testing.

Currently the tests are very limited - only covering two functions, and
not covering the multitude of Vim/Neovim versions currently available
in the container.

To run it on a machine having Docker installed, execute:

    ./test/run-tests