Fix doc comment code highlighting when indented
I forgot about these lines I removed in this commit when I improved the
doc comment code highlighting, and I thought I had made this sceenario
work properly. Ah well, it works now, so meh.

I removed the rustCommentLine rule because it was just wrong; given this
Rust code:

    /// ```
    /// foo
    // bar
    /// ```

… the “// bar” line is a regular comment rather than Rust code inside
a code fence, so we should leave it to just be regularly highlighted.

I replaced the rustCommentLineDoc rule with rustCommentLineDocLeader so
that it could only be activated inside `///` comments and not in `/**`
comments—though nesting doc comment blocks is definitely unusual!

Fixes #239.