rustfmt: Avoid the use of writefile() and expand('%')
Similarly to what other reformatters plugins do, we should use a
combination of setline() and line deletion command to carefully
replace the lines of the buffer(). There's issue #236 that seems
to not regress in this fix - the problem with '1,$d _' is that it
deleted the entire buffer. We should only delete for line numbers
that beyond the newly applied buffer content.

This has a side effect of handling #260 and #259, as expand('%')
is not relied upon.