RustFmt: fix non-existant temp file
In a Windows Vim, the configured shell, which may be simply CMD.EXE,
does not redirect stderr to the temp file, and thus the tempfile is not
crated, and we get an error about the attempt to delete it later.

In the future, we will use Vim's job_start API for differentiating
between stderr and stdout in a portable and precise way. For now, avoid
the annoying error when invoking RustFmt.

Fixes #285.