Restore cwd on rustfmt failure before opening lwindow
If rustfmt fails RunRustFmt helpfully opens the lwindow to show
the errors rustfmt encountered.
The 'lwindow' command opens the location list window, and makes that
the active vim window.

Before rustfmt is executed RunRustFmt change CWD to the directory
where the file to be formated resides. CWD is restored before exiting
RunRustFmt.

The problem was that location list window was opened before restoring
CWD, which meant that we only restored the directory of the lwindow.
The window holding the formatted rust file is left with CWD of that
file.

The obvious solution is to restore the CWD *before* opening the
location list window.