Allow attributes to contain nested brackets.
With Rust 1.34, proc-macro attributes now accept arbitrary tokens as arguments, allowing a user to write an attribute e.g. `#[my_attribute(a = [1, 2, 3], b = ("x", "y", "z")]`
The previous attribute highlighter would stop at the first `]`, leaving a lot of the attribute unhighlighted. This patch changes the highlighter to be more robust to nested brackets and parentheses.

As before, only derives, comments, and strings get special formatting inside of an attribute. It may make more sense to allow arbitrary syntax highlighting inside of the attribute arguments, but that's outside the scope of this patch.

Fixes #320