Don't change user's formatoptions
I've hit a bug in which hitting enter it will insert a comment start, any word that overflowed the previous line and set the cursor to the start of the line (before the comment start).

Which result in some amazing stuff like:

```rust 
// Some awesome long
ent. // comm
```

(It should say "Some awesome long comment.")

This change remove setting the formatoptions for the user, if you use Vim you can decide this for yourself.