perf(syntax): use display where applicable
One of the reasons that user type highlighting was rejected initially
was that it "slowed down" the syntax engine.

While I cannot reproduce this on my ancient pre-2010 Macbook, I am
open to the idea that it may happen.

Using `display` in the cases where it *can* be used will allow
performance improvement and may negate some of the additional overhead
imposed by the previous commits in this branch.