ref(syntax): add @rustComments for comments
There was some repetition of which comment groups were `contained` in
folds, and which weren't. Some even dictated inclusion of nested groups,
which seem to be intended for nesting within a `contained` top-level
comment.

This commit adds a dedicated `cluster` for including the top-level
`rustComment`* groups, and replaces usages where applicable.