feat(syntax): diff between anonymous lifetimes
This commit makes it so that regular lifetimes are highlighted as
`Label`s, and anonymous lifetimes are highlighted as `Special`. This is
because of our current distinction for `'special`, which is a
keyword-lifetime that has special meaning.

I felt that we should provide specific highlighting for all of our
lifetimes with special meanings.