feat(syntax): use PreCondit for cfg + cfg_attr
The `PreCondit` highlight group is meant to be used for conditional
compilation. See `:h group-name` for more info.

I'm trying to keep the changing of groups to a minimum in this PR,
though a cursory look over that help page will probably be helpful in
the future when going back and reviewing some of the choices which have
been made here.