fix(syntax): `pub(in crate)`: `in` not highlighted
This is because we were using a `match` instead of a `region`. The
`region` allows us to specify a `matchgroup` without explicitly
detailing that the `matchgroup` is a part of the region.

In our case, the `rustNoise` gives a `nextgroup` of `@rustIdentifiers`,
so `in` was being picked up as an identifier. No, the `rustNoise` is not
actually part of the match, so the `nextgroup` doesn't apply.