fix(syntax): make rustFuncName definition more correct
Previously we were allowing either `::` or `<.*>` to define the
`rustFuncName`. This is not correct: if `::` is present, `<.*>` must be
present also. However, `<.*>` doesn't need to be present, so it is
optional. We wrap these two together under one `?` group and `::` under
a further `?` group so that the above definition holds.