Fix indentation for closing parentheses
To help explain this problem, consider this input (where | is the
cursor):

    foo(|)

Prior to this commit, pressing Enter would result in the following:

    foo(
        |)

That is, the closing parenthesis is indented by one level. This is the
case regardless of the cindent options/keys chosen.

To fix this, we manually indent lines that start with a ")" (ignoring
leading whitespace). Such lines are indented according to the
indentation of the line that contained the opening parenthesis.

Fixes #443