fix(syntax): keywords detected as `rustIdentifier`
A side effect of yesterdays changes. While creating a demo file, I found
more issues. To fix them, I created a `@rustTokens` group with a correct
ordering for detecting the type of any given token when the context is
too generic to know for sure what the next token will be.