fix(syntax): `rustFoldBraces` region not ending
This issue was caused by the braces (`{`, `}`) being part of the
`rustNoise` group _and_ `rustFoldBraces`. To fix it, I had to remove the
braces from `rustNoise` and add `contains` arguments to `rustFoldBraces`
to maintain a fix from b2c540a.

There may be more groups that have to be in the `contains` arguments,
but they will have to be added as discovered.

RELATED: https://github.com/rust-lang/rust.vim/pull/431#issuecomment-900799162