ref(syntax): link `rustStructure` to `Keyword`
Some colorschemes don't distinguish between `Structure` and `Type`, so
this leads to confusion when `struct` and `enum` keywords are shown as
builtin types. Plus, these keywords aren't actually structures
themselves so I suppose that it is better for them to be highlighted as
a keyword anyway.

Thanks @Dispersia and @phi-gamma for pointing this out!