Allow option to highlight operator `as` as keyword
While `as` is functionally an operator, it is often highlighted as if it
were a keyword. This change gives the user the option to link
`rustOperatorKeyword` to `Keyword` while preserving the current default
behavior. This allows for `as` to be highlighted like a keyword
separately from other operators.

The documentation lists `as` as a keyword and consistently highlights it
as such.

https://doc.rust-lang.org/std/index.html#keywords
https://doc.rust-lang.org/std/keyword.as.html